+++ pkg/topology/scheduler/autoplacetopology/autoplacetopology.go
@@ -3,6 +3,7 @@ package autoplacetopology
 import (
 	"context"
 	"fmt"
+	"sort"
 
 	linstor "github.com/LINBIT/golinstor"
 	lapi "github.com/LINBIT/golinstor/client"
@@ -51,49 +52,51 @@ func (s *Scheduler) Create(ctx context.Context, volId string, params *volume.Par
 	log.WithField("requirements", topologies).Trace("got topology requirement")
 
 	for _, preferred := range topologies.GetPreferred() {
-		log := log.WithField("segments", preferred.GetSegments())
-
-		nodes, err := s.NodesForTopology(ctx, preferred.GetSegments())
+		err := s.PlaceAccessibleToSegment(ctx, volId, preferred.GetSegments(), params.AllowRemoteVolumeAccess)
 		if err != nil {
-			return fmt.Errorf("failed to get preferred node list from segments: %w", err)
-		}
-
-		log.WithField("nodes", nodes).Trace("try initial placement on preferred nodes")
-
-		apRequest := lapi.AutoPlaceRequest{SelectFilter: lapi.AutoSelectFilter{NodeNameList: nodes}}
-
-		if len(nodes) < int(params.PlacementCount) {
-			apRequest.SelectFilter.PlaceCount = int32(len(nodes))
-		}
-
-		err = s.Resources.Autoplace(ctx, volId, apRequest)
-		if err != nil {
-			log.WithError(err).Trace("failed to autoplace")
+			log.WithError(err).Debug("failed to place on preferred segment")
 		} else {
-			log.Trace("successfully placed on preferred node")
+			log.Debug("placed accessible to preferred segment")
 			break
 		}
 	}
 
 	// By now we should have placed a volume on one of the preferred nodes (or there were no preferred nodes). Now
 	// we can try autoplacing the rest. Initially, we want to restrict ourselves to just the requisite nodes.
-	var requisiteNodes []string
+	// However, we _can_ always expand to the accessible volumes as per remote volume policy.
+	requisiteNodes := make(map[string]struct{})
 
 	for _, requisite := range topologies.GetRequisite() {
-		nodes, err := s.NodesForTopology(ctx, requisite.GetSegments())
-		if err != nil {
-			return fmt.Errorf("failed to get preferred node list from segments: %w", err)
-		}
+		for _, seg := range params.AllowRemoteVolumeAccess.AccessibleSegments(requisite.GetSegments()) {
+			log := log.WithField("segments", seg)
 
-		requisiteNodes = append(requisiteNodes, nodes...)
+			nodes, err := s.NodesForTopology(ctx, seg)
+			if err != nil {
+				return fmt.Errorf("failed to get preferred node list from segments: %w", err)
+			}
+
+			log.WithField("nodes", nodes).Trace("got nodes for segment")
+
+			for i := range nodes {
+				requisiteNodes[nodes[i]] = struct{}{}
+			}
+		}
 	}
 
 	log.WithField("requisite", requisiteNodes).Trace("got requisite nodes")
 
 	if len(requisiteNodes) > 0 {
+		requisiteNodesList := make([]string, 0, len(requisiteNodes))
+		for k := range requisiteNodes {
+			requisiteNodesList = append(requisiteNodesList, k)
+		}
+
+		// Sort, for testing
+		sort.Strings(requisiteNodesList)
+
 		// We do have requisite nodes, so we need to autoplace just on those nodes.
 		req := lapi.AutoPlaceRequest{
-			SelectFilter: lapi.AutoSelectFilter{NodeNameList: requisiteNodes},
+			SelectFilter: lapi.AutoSelectFilter{NodeNameList: requisiteNodesList},
 		}
 
 		// We might need to restrict autoplace here. We could have just one requisite node, but a placement count of 3.
@@ -132,8 +135,66 @@ func (s *Scheduler) Create(ctx context.Context, volId string, params *volume.Par
 	return nil
 }
 
-func (s *Scheduler) AccessibleTopologies(ctx context.Context, volId string, allowDisklessAccess bool) ([]*csi.Topology, error) {
-	return s.GenericAccessibleTopologies(ctx, volId, allowDisklessAccess)
+// PlaceAccessibleToSegment tries to place a replica accessible to the given segment.
+//
+// Initially, placement on an exactly matching node is tried. If that is not possible, the remoteAccessPolicy is
+// used to determine other nodes that would grant the given segment access.
+func (s *Scheduler) PlaceAccessibleToSegment(ctx context.Context, volId string, segments map[string]string, remoteAccessPolicy volume.RemoteAccessPolicy) error {
+	log := s.log.WithField("volume", volId).WithField("segments", segments)
+
+	nodes, err := s.NodesForTopology(ctx, segments)
+	if err != nil {
+		return fmt.Errorf("failed to get preferred node list from segments: %w", err)
+	}
+
+	log.WithField("nodes", nodes).Trace("try initial placement on preferred nodes")
+
+	apRequest := lapi.AutoPlaceRequest{SelectFilter: lapi.AutoSelectFilter{NodeNameList: nodes, PlaceCount: 1}}
+
+	err = s.Resources.Autoplace(ctx, volId, apRequest)
+	if err != nil {
+		log.WithError(err).Trace("failed to autoplace")
+	} else {
+		log.Trace("successfully placed on preferred node")
+		return nil
+	}
+
+	log.Trace("exact segments match failed, try with remote access policy")
+
+	accessibleSegments := remoteAccessPolicy.AccessibleSegments(segments)
+
+	log.WithField("accessibleSegments", accessibleSegments).Trace("got accessible segments")
+
+	for _, seg := range accessibleSegments {
+		log := log.WithField("segments", seg)
+
+		nodes, err := s.NodesForTopology(ctx, seg)
+		if err != nil {
+			log.WithError(err).Warn("failed to get preferred node list from segments")
+			continue
+		}
+
+		if len(nodes) == 0 {
+			log.Trace("no matching node")
+			continue
+		}
+
+		apRequest := lapi.AutoPlaceRequest{SelectFilter: lapi.AutoSelectFilter{NodeNameList: nodes, PlaceCount: 1}}
+
+		err = s.Resources.Autoplace(ctx, volId, apRequest)
+		if err != nil {
+			log.WithError(err).Trace("failed to autoplace")
+		} else {
+			log.Trace("successfully placed reachable for preferred node")
+			return nil
+		}
+	}
+
+	return fmt.Errorf("failed to place resource")
+}
+
+func (s *Scheduler) AccessibleTopologies(ctx context.Context, volId string, remoteAccessPolicy volume.RemoteAccessPolicy) ([]*csi.Topology, error) {
+	return s.GenericAccessibleTopologies(ctx, volId, remoteAccessPolicy)
 }
 
 func NewScheduler(c *lc.HighLevelClient, l *logrus.Entry) *Scheduler {