+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `allowRemoteVolumeAccess` supports new values, allowing a more fine-grained controlled on diskless access.
+
 ## [0.16.1] - 2021-11-08
 
 ### Changed